Refactor Simple Client into Simple Screen
This Design pattern will be useful for developing multiple screens in the annotation tool
