+++ client/definitions.py
@@ -1,0 +1,4 @@
+import os
+
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
+SCREEN_DIR = os.path.join(ROOT_DIR, 'screens')
