+++ client/screens/simple_screen.kv
@@ -1,6 +1,3 @@
-<MyScreenManager>:
-    SimpleScreen:
-
 <SimpleScreen>:
     status_label: status_label
     BoxLayout:
