+++ client/screens/simple_screen.py
@@ -1,0 +1,24 @@
+import os
+
+from kivy.lang import Builder
+from kivy.properties import ObjectProperty
+from kivy.network.urlrequest import UrlRequest
+from kivy.uix.screenmanager import Screen
+
+from client.client_config import ClientConfig
+from client.definitions import SCREEN_DIR
+
+
+# Load corresponding kivy file
+Builder.load_file(os.path.join(SCREEN_DIR, 'simple_screen.kv'))
+
+
+class SimpleScreen(Screen):
+    # Placeholder for label generated by .kv file
+    status_label = ObjectProperty(None)
+
+    def heartbeat_request(self, *args):
+        UrlRequest(ClientConfig.SERVER_URL, self.handle_heartbeat)
+
+    def handle_heartbeat(self, req, result):
+        self.status_label.text = str("Last Heartbeat: " + str(result))
