+++ client/simple_client.py
@@ -1,28 +1,25 @@
 from kivy.app import App
-from kivy.properties import ObjectProperty
-from kivy.network.urlrequest import UrlRequest
-from kivy.uix.screenmanager import ScreenManager, Screen
+from kivy.uix.screenmanager import ScreenManager
 
-from client.client_config import ClientConfig
+from client.screens.simple_screen import SimpleScreen
 
 
 class MyScreenManager(ScreenManager):
-    pass
+    """
+    Defines the screens which should be included in the ScreenManager at launch.
+    Use the ScreenManager to handle transitions between Screens.
+    """
 
-
-class SimpleScreen(Screen):
-    client_config = ClientConfig()
-    # Placeholder for label generated by .kv file
-    status_label = ObjectProperty(None)
-
-    def heartbeat_request(self, *args):
-        UrlRequest(self.client_config.SERVER_URL, self.handle_heartbeat)
-
-    def handle_heartbeat(self, req, result):
-        self.status_label.text = str("Last Heartbeat: " + str(result))
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.add_widget(SimpleScreen())
 
 
 class SimpleClientApp(App):
+    """
+    The launch point for the application itself.
+    """
+
     def build(self):
         return MyScreenManager()
